Add Remove Cycle check for owning

+++ src/main/java/com/csl456/bikerentalapp/resources/CycleResource.java
@@ -1,33 +1,32 @@
 package com.csl456.bikerentalapp.resources;
 
-import java.util.List;
-import java.util.Optional;
-
-import javax.ws.rs.Consumes;
-import javax.ws.rs.DELETE;
-import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.QueryParam;
-import javax.ws.rs.core.MediaType;
-
 import com.csl456.bikerentalapp.core.Cycle;
 import com.csl456.bikerentalapp.core.UserRole;
 import com.csl456.bikerentalapp.db.CycleDAO;
+import com.csl456.bikerentalapp.db.SessionDAO;
+import com.csl456.bikerentalapp.db.UserDAO;
 import com.csl456.bikerentalapp.filter.RolesAllowed;
-
 import io.dropwizard.hibernate.UnitOfWork;
+
+import javax.ws.rs.*;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
+import java.util.List;
+import java.util.Optional;
 
 @Path("/cycle")
 @Consumes(MediaType.APPLICATION_JSON)
 @Produces(MediaType.APPLICATION_JSON)
 public class CycleResource {
 	private final CycleDAO cycleDAO;
+	private final UserDAO userDAO;
+	private final SessionDAO sessionDAO;
 
-	public CycleResource(CycleDAO cycleDAO) {
+	public CycleResource(CycleDAO cycleDAO, UserDAO userDAO,
+						 SessionDAO sessionDAO) {
 		this.cycleDAO = cycleDAO;
+		this.userDAO = userDAO;
+		this.sessionDAO = sessionDAO;
 	}
 
 	@POST
@@ -59,7 +58,12 @@ public class CycleResource {
 	@UnitOfWork
 	@Path("{id}")
 	@RolesAllowed(UserRole.ADMIN)
-	public Cycle removeCycle(@PathParam("id") int id) {
+	public Cycle removeCycle(@PathParam("id") int id, @HeaderParam(
+			"Access_Token") String accessToken) {
+		if (cycleDAO.findById(id).getOwnerId() != userDAO.findByUserName(sessionDAO.getUserName(accessToken)).getPersonId()) {
+			throw new WebApplicationException("You do not own this cycle",
+					Response.Status.UNAUTHORIZED);
+		}
 		return cycleDAO.remove(id);
 	}
